# docs/api-docs.yaml
openapi: 3.1.0
info:
  title: EcoMóstoles B2B API
  description: API REST para el intercambio de residuos
  version: "1.0"
servers:
- url: https://localhost:8443
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/v1/agreements/{id}:
    get:
      tags:
      - Agreements
      summary: Get agreement by ID
      description: Retrieves the detail of a single agreement.
      operationId: getAgreementById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Agreement found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgreementDTO"
        "404":
          description: Agreement not found
        "500":
          description: Unexpected server error
    put:
      tags:
      - Agreements
      summary: Update an agreement
      description: "Updates fields or changes the state of an agreement (e.g., to\
        \ COMPLETED)."
      operationId: updateAgreement
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgreementDTO"
        required: true
      responses:
        "200":
          description: Agreement updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgreementDTO"
        "400":
          description: Validation failed
        "404":
          description: Agreement not found
    delete:
      tags:
      - Agreements
      summary: Delete an agreement
      description: "Cancels or deletes an agreement, if allowed."
      operationId: deleteAgreement
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Agreement deleted successfully
        "400":
          description: Cannot delete agreement in its current state
        "404":
          description: Agreement not found
  /api/v1/agreements/{id}/status:
    put:
      tags:
      - Agreements
      summary: Update agreement status
      description: Updates only the status of an existing agreement.
      operationId: updateAgreementStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                type: string
        required: true
      responses:
        "200":
          description: Status updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgreementDTO"
        "404":
          description: Agreement not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgreementDTO"
...
  /api/v1/agreements:
    get:
      tags:
      - Agreements
      summary: List agreements
      description: Returns a paginated list of commercial agreements.
      operationId: getAllAgreements
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 12
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "registrationDate,DESC"
          items:
            type: string
      responses:
        "200":
          description: Page of agreements returned successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgreementDTO"
        "500":
          description: Unexpected server error
    post:
      tags:
      - Agreements
      summary: Create an agreement
      description: Proposes a new commercial agreement.
      operationId: createAgreement
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgreementDTO"
        required: true
      responses:
        "201":
          description: Agreement created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgreementDTO"
        "400":
          description: Validation failed
...
components:
  schemas:
    CompanyDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        commercialName:
          type: string
          minLength: 1
        contactEmail:
          type: string
          minLength: 1
        taxId:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        industrialSector:
          type: string
          minLength: 1
        description:
          type: string
          maxLength: 255
          minLength: 0
        role:
          type: string
        co2Saved:
          type: number
          format: double
        ranking:
          type: integer
          format: int32
        sector:
          type: string
        verified:
          type: boolean
        initial:
          type: string
        formattedCo2Saved:
          type: string
      required:
      - address
      - commercialName
      - contactEmail
      - description
      - industrialSector
      - phone
      - taxId
...
    AgreementDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        exchangedMaterial:
          type: string
        quantity:
          type: number
          format: double
        unit:
          type: string
        agreedPrice:
          type: number
          format: double
        pickupDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - ACCEPTED
          - REJECTED
        notes:
          type: string
        registrationDate:
          type: string
          format: date-time
        originCompany:
          $ref: "#/components/schemas/CompanyDTO"
        destinationCompany:
          $ref: "#/components/schemas/CompanyDTO"
        offerId:
          type: integer
          format: int64
        demandId:
          type: integer
          format: int64
        platformCommission:
          type: number
          format: double
        co2Impact:
          type: number
          format: double
        formattedRegistrationDate:
          type: string
        formattedCo2Impact:
          type: string
        deletable:
          type: boolean
        formattedPlatformCommission:
          type: string
        formattedAgreedPrice:
          type: string
        formattedQuantity:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Introduce el token JWT obtenido en /api/v1/auth/login
      in: header
      scheme: bearer
      bearerFormat: JWT